backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
easyocr
ultralytics
openai
python-dotenv
